requirements.txt
numpy
tqdm
Pillow
shapely
